feat: sort findMany actions
