+++ packages/core/content-releases/admin/src/components/CMReleasesContainer.tsx
@@ -68,6 +68,7 @@ const AddActionToReleaseModal = ({
   const { formatMessage } = useIntl();
   const toggleNotification = useNotification();
   const { formatAPIError } = useAPIErrorHandler();
+  const { modifiedData } = useCMEditViewDataManager();
 
   // Get all 'pending' releases that do not have the entry attached
   const response = useGetReleasesForEntryQuery({
@@ -80,9 +81,11 @@ const AddActionToReleaseModal = ({
   const [createReleaseAction, { isLoading }] = useCreateReleaseActionMutation();
 
   const handleSubmit = async (values: FormValues) => {
+    const locale = modifiedData.locale as string | undefined;
     const releaseActionEntry = {
       contentType: contentTypeUid,
       id: entryId,
+      locale,
     };
     const response = await createReleaseAction({
       body: { type: values.type, entry: releaseActionEntry },
